# api/docs/swagger.yaml
swagger: '2.0'
info:
  title: cheqd Network - REST and gRPC Gateway Documentation
  description: Documentation describing REST and gRPC Gateway endpoints for cheqd network.
  version: v1.x
  contact:
    name: Cheqd Foundation Limited
    url: https://cheqd.io
  license:
    name: Apache 2.0
    url: https://github.com/cheqd/cheqd-node/blob/main/LICENSE
basePath: /swagger/
schemes:
  - http
  - https
consumes:
  - application/json
  - application/json+ld
  - '*/*'
produces:
  - application/json
  - application/json+ld
  - '*/*'
...
paths:
...
  /cheqd/resource/v2/{collection_id}/metadata:
    get:
      summary: Fetch all resource metadata from a collection with a given collection_id
      operationId: CollectionResources
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              resources:
                type: array
                items:
                  type: object
                  properties:
                    collection_id:
                      type: string
                      title: >-
                        collection_id is the ID of the collection that the
                        Resource belongs to. Defined client-side.

                        This field is the unique identifier of the DID linked to
                        this Resource
                    id:
                      type: string
                      title: >-
                        id is the ID of the Resource. Defined client-side.

                        This field is a unique identifier for this specific
                        version of the Resource.

                        Format: <uuid>
                    name:
                      type: string
                      title: >-
                        name is a human-readable name for the Resource. Defined
                        client-side.

                        Does not change between different versions.

                        Example: PassportSchema, EducationTrustRegistry
                    version:
                      type: string
                      title: >-
                        version is a human-readable semantic version for the
                        Resource. Defined client-side.

                        Stored as a string. OPTIONAL.

                        Example: 1.0.0, v2.1.0
                    resource_type:
                      type: string
                      title: >-
                        resource_type is a Resource type that identifies what
                        the Resource is. Defined client-side.

                        This is NOT the same as the resource's media type.

                        Example: AnonCredsSchema, StatusList2021
                    also_known_as:
                      type: array
                      items:
                        type: object
                        properties:
                          uri:
                            type: string
                            title: >-
                              uri is the URI of the Resource.

                              Examples:

                              -
                              did:canow:testnet:MjYxNzYKMjYxNzYK/resources/4600ea35-8916-4ac4-b412-55b8f49dd94e

                              -
                              https://resolver..cheqd.net/1.0/identifiers/did:canow:testnet:MjYxNzYKMjYxNzYK/resources/4600ea35-8916-4ac4-b412-55b8f49dd94e

                              - https://example.com/example.json

                              -
                              https://gateway.ipfs.io/ipfs/bafybeihetj2ng3d74k7t754atv2s5dk76pcqtvxls6dntef3xa6rax25xe

                              -
                              ipfs://bafybeihetj2ng3d74k7t754atv2s5dk76pcqtvxls6dntef3xa6rax25xe
                          description:
                            type: string
                            title: >-
                              description is a human-readable description of the
                              URI. Defined client-side.

                              Examples:

                              - did-uri

                              - http-uri

                              - ipfs-uri
                        description: >-
                          AlternativeUri are alternative URIs that can be used
                          to access the Resource.

                          By default, at least the DID URI equivalent of the
                          Resource is populated.
                      description: List of alternative URIs for the SAME Resource.
                    media_type:
                      type: string
                      title: >-
                        media_type is IANA media type of the Resource. Defined
                        ledger-side.

                        Example: application/json, image/png
                    created:
                      type: string
                      format: date-time
                      title: >-
                        created is the time at which the Resource was created.
                        Defined ledger-side.

                        Format: RFC3339

                        Example: 2021-01-01T00:00:00Z
                    checksum:
                      type: string
                      title: >-
                        checksum is a SHA-256 checksum hash of the Resource.
                        Defined ledger-side.

                        Example:
                        d14a028c2a3a2bc9476102bb288234c415a2b01f828ea62ac5b3e42f
                    previous_version_id:
                      type: string
                      title: >-
                        previous_version_id is the ID of the previous version of
                        the Resource. Defined ledger-side.

                        This is based on the Resource's name and Resource type
                        to determine whether it's the same Resource.

                        Format: <uuid>
                    next_version_id:
                      type: string
                      title: >-
                        next_version_id is the ID of the next version of the
                        Resource. Defined ledger-side.

                        This is based on the Resource's name and Resource type
                        to determine whether it's the same Resource.

                        Format: <uuid>
                  title: Metadata stores the metadata of a DID-Linked Resource
                title: resources is the requested collection of resource metadata
            title: >-
              QueryCollectionResourcesResponse is the response type for the
              Query/CollectionResources RPC method
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: collection_id
          description: >-
            collection_id is an identifier of the DidDocument the resource
            belongs to.

            Format: <unique-identifier>


            Examples:

            - c82f2b02-bdab-4dd7-b833-3e143745d612

            - wGHEXrZvJxR8vw5P3UWH1j
          in: path
          required: true
          type: string
      tags:
        - Query
  /cheqd/resource/v2/{collection_id}/resources/{id}:
    get:
      summary: Fetch a resource from a collection with a given collection_id and id
      operationId: Resource
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              resource:
                title: >-
                  Successful resolution of the resource returns the following:

                  - resource is the requested resource

                  - metadata is the resource metadata associated with the
                  requested resource
                type: object
                properties:
                  resource:
                    type: object
                    properties:
                      data:
                        type: string
                        format: byte
                        title: bytes is the raw data of the Resource
                    title: Resource stores the contents of a DID-Linked Resource
                  metadata:
                    type: object
                    properties:
                      collection_id:
                        type: string
                        title: >-
                          collection_id is the ID of the collection that the
                          Resource belongs to. Defined client-side.

                          This field is the unique identifier of the DID linked
                          to this Resource
                      id:
                        type: string
                        title: >-
                          id is the ID of the Resource. Defined client-side.

                          This field is a unique identifier for this specific
                          version of the Resource.

                          Format: <uuid>
                      name:
                        type: string
                        title: >-
                          name is a human-readable name for the Resource.
                          Defined client-side.

                          Does not change between different versions.

                          Example: PassportSchema, EducationTrustRegistry
                      version:
                        type: string
                        title: >-
                          version is a human-readable semantic version for the
                          Resource. Defined client-side.

                          Stored as a string. OPTIONAL.

                          Example: 1.0.0, v2.1.0
                      resource_type:
                        type: string
                        title: >-
                          resource_type is a Resource type that identifies what
                          the Resource is. Defined client-side.

                          This is NOT the same as the resource's media type.

                          Example: AnonCredsSchema, StatusList2021
                      also_known_as:
                        type: array
                        items:
                          type: object
                          properties:
                            uri:
                              type: string
                              title: >-
                                uri is the URI of the Resource.

                                Examples:

                                -
                                did:canow:testnet:MjYxNzYKMjYxNzYK/resources/4600ea35-8916-4ac4-b412-55b8f49dd94e

                                -
                                https://resolver..cheqd.net/1.0/identifiers/did:canow:testnet:MjYxNzYKMjYxNzYK/resources/4600ea35-8916-4ac4-b412-55b8f49dd94e

                                - https://example.com/example.json

                                -
                                https://gateway.ipfs.io/ipfs/bafybeihetj2ng3d74k7t754atv2s5dk76pcqtvxls6dntef3xa6rax25xe

                                -
                                ipfs://bafybeihetj2ng3d74k7t754atv2s5dk76pcqtvxls6dntef3xa6rax25xe
                            description:
                              type: string
                              title: >-
                                description is a human-readable description of
                                the URI. Defined client-side.

                                Examples:

                                - did-uri

                                - http-uri

                                - ipfs-uri
                          description: >-
                            AlternativeUri are alternative URIs that can be used
                            to access the Resource.

                            By default, at least the DID URI equivalent of the
                            Resource is populated.
                        description: List of alternative URIs for the SAME Resource.
                      media_type:
                        type: string
                        title: >-
                          media_type is IANA media type of the Resource. Defined
                          ledger-side.

                          Example: application/json, image/png
                      created:
                        type: string
                        format: date-time
                        title: >-
                          created is the time at which the Resource was created.
                          Defined ledger-side.

                          Format: RFC3339

                          Example: 2021-01-01T00:00:00Z
                      checksum:
                        type: string
                        title: >-
                          checksum is a SHA-256 checksum hash of the Resource.
                          Defined ledger-side.

                          Example:
                          d14a028c2a3a2bc9476102bb288234c415a2b01f828ea62ac5b3e42f
                      previous_version_id:
                        type: string
                        title: >-
                          previous_version_id is the ID of the previous version
                          of the Resource. Defined ledger-side.

                          This is based on the Resource's name and Resource type
                          to determine whether it's the same Resource.

                          Format: <uuid>
                      next_version_id:
                        type: string
                        title: >-
                          next_version_id is the ID of the next version of the
                          Resource. Defined ledger-side.

                          This is based on the Resource's name and Resource type
                          to determine whether it's the same Resource.

                          Format: <uuid>
                    title: Metadata stores the metadata of a DID-Linked Resource
            title: >-
              QueryResourceResponse is the response type for the Query/Resource
              RPC method
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: collection_id
          description: >-
            collection_id is an identifier of the DidDocument the resource
            belongs to.

            Format: <unique-identifier>


            Examples:

            - c82f2b02-bdab-4dd7-b833-3e143745d612

            - wGHEXrZvJxR8vw5P3UWH1j
          in: path
          required: true
          type: string
        - name: id
          description: |-
            id is a unique id of the resource.
            Format: <uuid>
          in: path
          required: true
          type: string
      tags:
        - Query
...